perf: Ajustando lógica na localização para melhor organização e apagado um arquivo desnecessário

+++ controllers/locationController.js
@@ -6,7 +6,7 @@ export const getAll = async (req, res) => {
     res.status(200).json(getLocation);
   } catch (error) {
     console.error("Erro ao buscar as localizações:", error);
-    res.status(500).json({ message: "Erro interno do servidor." });
+    res.status(500).json({ message: "Erro interno do servidor: ", error });
   }
 };
 
@@ -14,8 +14,19 @@ export const createLocation = async (req, res) => {
   try {
     const { name, description } = req.body;
 
-    if (!name) {
-      return res.status(400).json({ message: "Campos obrigatórios ausentes." });
+    const requiredFields = ["name"];
+    const missingFields = [];
+
+    requiredFields.forEach((field) => {
+      if (!req.body[field]) {
+        missingFields.push(field);
+      }
+    });
+
+    if (missingFields.length > 0) {
+      return res.status(400).json({
+        error: `Campos obrigatórios ausentes: ${missingFields.join(", ")}`,
+      });
     }
 
     const location = {
@@ -23,11 +34,11 @@ export const createLocation = async (req, res) => {
       description,
     };
 
-    const response = await Location.createLocation(location);
+    await Location.createLocation(location);
     res.status(201).json({ msg: "Localização registrada com sucesso!" });
   } catch (error) {
-    console.error("Erro ao criar localização:", error);
-    res.status(500).json("Erro interno do servidor:", error);
+    console.error("Erro ao criar localização: ", error);
+    res.status(500).json("Erro interno do servidor: ", error);
   }
 };
 
@@ -41,25 +52,33 @@ export const updateLocation = async (req, res) => {
       description,
       active,
     };
-    // const location = await Location.findById(id).exec();
-    // if (!location) {
-    //   res.status(204).json({ msg: `Nenhuma localização encontrada com esse ID!` });
-    // }
+
+    const locationExists = await Location.getLocationById(id);
+    if (!locationExists) {
+      return res.status(404).json({
+        error: "Nenhuma localização encontrada! Forneça um ID válido.",
+      });
+    }
+
     await Location.updateLocation(id, response);
     res.status(201).json({ msg: "Localização atualizada com sucesso!" });
   } catch (error) {
-    console.log(error);
-    res.status(500).json({ error: "Erro interno no servidor." });
+    console.log("Erro ao atualizar a localização: ", error);
+    res.status(500).json({ error: "Erro interno no servidor: ", error });
   }
 };
 
 export const deleteLocation = async (req, res) => {
   try {
     const { id } = req.params;
-    // const location = await Location.findById(id).exec();
-    // if (!location) {
-    //   res.status(204).json({ msg: `Nenhuma localização encontrada!` });
-    // }
+
+    const locationExists = await Location.getLocationById(id);
+    if (!locationExists) {
+      return res.status(404).json({
+        error: "Nenhuma localização encontrada! Forneça um ID válido.",
+      });
+    }
+
     await Location.deleteLocation(id);
     res.status(200).json({ msg: "Localização deletada com sucesso!" });
   } catch (error) {
